Move download function to background
